Extracted the Surrounding Edges Logic
